pyproject.toml
[tool.poetry]
name = "root"
version = "0.0.0"
description=""
authors=["srp-31 <srp@tier.app>"]


[tool.poetry.dependencies]
python = ">=3.6.1,<4.0"
argparse = "1.1"
imutils = "0.5.3"
keras = "2.3.1"
matplotlib = "3.2.1"
numpy = "1.18.2"
opencv-python = "==4.2.0.*"
pillow = "7.2.0"
scikit-learn = "0.23.1"
scipy = "1.4.1"
tensorflow = ">=1.15.2"
wheel = "^0.36.2"
seldon-core = "^1.5.1"
Flask = "^1.1.2"
gunicorn = "^20.0.4"
dash = "^1.18.1"
dash-renderer = "^1.8.3"
mlrun = "^0.5.4"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"  # "*" does not work with black's versioning scheme
bump2version = "*"
coverage = "*"
flake8 = "*"
isort = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
tox = "*"
Sphinx = "*"
pydocstyle = "*"
sphinx_rtd_theme = "*"

[tool.black]
line-length = "99"
target-version = ['py36', 'py37', 'py38']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | docs
  )/
  | setup.py           # also separately exclude a file named foo.py in
                       # the root of the project
)
'''


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

